# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NetterTech Events REST API
  description: |
    REST API for the NetterTech Events WordPress plugin.

    ## Scope & Status

    This specification is being incrementally reconciled with the implementation in
    `includes/API/`. As of 2026-04-20, the core plugin ships exactly these 20 routes
    across 7 controllers (verified via `grep register_rest_route includes/API`):

    | Controller | Routes |
    |------------|--------|
    | EventsController | `GET /events/upcoming`, `GET /events/range`, `GET /events/{id}`, `GET /occurrences` |
    | ICalController | `GET /ical/feed`, `GET /ical/event/{id}`, `GET /ical/occurrence/{id}`, `POST /ical/import` |
    | WaitlistRestController | `POST /waitlist/join`, `GET /waitlist/status`, `POST /waitlist/leave` |
    | AttendeeFieldsController | `GET/POST /events/{event_id}/attendee-fields`, `GET/PATCH/DELETE /events/{event_id}/attendee-fields/{id}`, `GET /attendees/{attendee_id}/custom-fields` |
    | EventsAdminController | `GET/POST /admin/events`, `GET/PATCH/DELETE /admin/events/{id}` |
    | AttendeesAdminController | `GET/POST /admin/attendees`, `GET/PATCH/DELETE /admin/attendees/{id}` |
    | TicketTypesAdminController | `GET/POST /admin/ticket-types`, `GET/PATCH/DELETE /admin/ticket-types/{id}` |

    Routes listed below with the `Check-In` or `Check-In (Admin)` tag are documented
    here for historical reference but are **not registered by the core plugin** — they
...

...

    ## Authentication

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /admin/attendees:
    get:
      tags:
        - Attendees (Admin)
      summary: List attendees for an occurrence
      operationId: getAttendees
      security:
        - WordPressCookie: []
      parameters:
        - name: occurrence_id
          in: query
          required: true
          description: Occurrence ID (required)
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: ['', pending, confirmed, cancelled, refunded]
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
      responses:
        '200':
          description: List of attendees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attendee'
        '400':
          description: Missing occurrence_id
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Attendees (Admin)
      summary: Create an attendee
      operationId: createAttendee
      security:
        - WordPressCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendeeInput'
      responses:
        '201':
          description: Attendee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendee'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/attendees/{id}:
    get:
      tags:
        - Attendees (Admin)
      summary: Get a single attendee
      operationId: getAttendee
      security:
        - WordPressCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Attendee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendee'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Attendees (Admin)
      summary: Update an attendee
      operationId: updateAttendee
      security:
        - WordPressCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendeeInput'
      responses:
        '200':
          description: Attendee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendee'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Attendees (Admin)
      summary: Delete an attendee
      operationId: deleteAttendee
      security:
        - WordPressCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Attendee deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    WordPressCookie:
      type: apiKey
      in: cookie
      name: wordpress_logged_in
      description: WordPress authentication cookie
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        data:
          type: object
          properties:
            status:
              type: integer
              description: HTTP status code
...
    Attendee:
      type: object
      properties:
        id:
          type: integer
        occurrence_id:
          type: integer
        ticket_type_id:
          type: integer
          nullable: true
        wc_order_id:
          type: integer
          nullable: true
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        quantity:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [pending, confirmed, cancelled, refunded]
        checked_in:
          type: boolean
        checked_in_count:
          type: integer
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        accessibility_notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        _links:
          type: object
    AttendeeInput:
      type: object
      required:
        - occurrence_id
        - name
        - email
      properties:
        occurrence_id:
          type: integer
        ticket_type_id:
          type: integer
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        quantity:
          type: integer
          minimum: 1
          default: 1
        status:
          type: string
          enum: [pending, confirmed, cancelled, refunded]
          default: confirmed
        notes:
          type: string
        accessibility_notes:
          type: string
...